// src/interfaces/api.interface.ts
export interface baseRes<D = {}> {
	success: boolean
	code: number
	message: string
	data: D //返回的数据字段
}

export interface PageData<D = {}> {
	tempGroup: any
	bindList: any
	filePreviewUrl: string
	fileSaveUrl: string
	currentPage: number
	pageNum: number
	pageSize: number
	totalPage: number
	count: number
	list: D[]
	fileHost: string
	helpVideo: {
		videoUrl: string
		coverUrl: string
	}
}
// ...
export interface userData {
	enabled: any
	description: string
	permissionName: string
	id: string
	username: string
	realName: string
	nickname: string
	avatar: string
	telephone: string
	email: string
	sex: number
	birthday: number
	signature: string
	accessToken: string
	userOrigin: number
	auth: boolean
	lastLoginIp: string
	lastLoginDate: number
	createTs: number
	type: string
	isShowAllMenus: boolean
	roleList?: number[]
}
// ...
export interface platformsItem {
	key: number | string
	value: string
}
// ...
export interface userListItem {
	birthday: string
	email: string
	id: string
	idCard: string
	nickname: string
	realName: string
	roleList: null
	sex: number
	state: number
	telephone: string
	userOrigin: null
	username: string
	name?: string
}
// ...
export interface menusInfoData {
	id: number
	parentId: number
	menuName: string
	menuPath: string
	url: string
	icon: null
	sort: number
	level: number
	leaf: boolean
	type: number
	description: string
	children: menusInfoData[]
}

export interface changeMenus {
	id: number
	parentId: number
	menuName: string
	menuPath: string
	url: string
	icon: string
	sort: number
	level: number
	leaf: boolean
	type: number
	description: string
}

export interface rolesData<T> {
	userId?: string
	roleIds?: T[]
}

export interface rolesDataItem {
	ActualType: number
}

export interface rolesListItem {
	id: string
	code: string
	name: string
}

export interface comboItem {
	id: number
	name: string
	code: string
}

export interface rolesPermissionsItem {
	id: number
	permissionName: string
	description: string
	createTs: string
}

export interface menuTreeItem {
	children: menuTreeItem[]
	enabled: boolean
	hasPermission: boolean
	icon: string
	id: number
	leaf: boolean
	level: number
	menuName: string
	menuPath: string
	parentId: number
	sort: number
	title?: string
	type: number
	url: string
	key: number
}
// ...
export interface mealInfo {
	devId: string
	devPackageId: string
	expire: number
	lives: number
	onlineServices: number
	publishScenesExtra: number
	sceneSyncs: number
	version: number
	visitorInfos: number
}

export interface mealItem {
	id: string
	name: string
	publishScenes: number
	day: number
}

export interface mealList {
	packagesList: mealItem[]
	packageId: string
	devId: string
}

export interface tempParams {
	pageNum: number
	pageSize: number
	tempName?: string
	tempId?: string
	userId?: string
}
export interface devAddtemp {
	tempType: number
	businessId: number
	tempName: string
	createType: number
	userId: string
}
export interface devInfo {
	userId: string
	userName: string
	telephone: string
	tag: string
}
export interface UserTemp {
	pageNum?: number
	pageSize?: number
	tempName?: string
	tempId?: string
	phoneOrName?: string
	isCheck?: number
	syncStatic?: number
	serverType?: number
	startTime?: string
	endTime?: string
	devUserId: string
}
export interface tempInfo {
	validSz?: number
}
export interface accountOverview {
	delScenes: number
	importUsers: number
	livesUses: number
	onlineServicesUses: number
	publishScenesUses: number
	sceneSyncsUses: number
	totalScenes: number
	visitorInfosUses: number
	expireStr: string
	vrLooksUses: number
}
export interface developerPackagesInfo {
	totalScenes: number
	importUsers: number
	visitorInfos: number
	visitorInfosUses: number
	lives: number
	livesUses: number
	sceneSyncs: number
	sceneSyncsUses: number
	publishScenes: number
	publishScenesUses: number
	delScenes: number
	onlineServices: number
	onlineServicesUses: number
	expireStr: string
	vrLooksUses: number
	vrLooks: number
}

export interface resoureInfo {
	developerInfo: devInfo
	developerPackagesInfo: developerPackagesInfo
	accountOverview: accountOverview
}
// ...
export interface logListItem {
	list: logItem[]
}

export interface logItem {
	id: number
	requestUri: string
	method: string
	ip: string
	businessId: string
	businessTitle: string
	businessType: string
	businessDesc: string
	logDesc: string
	nickname: string
	creator: string
	createTs: string
}