// src/interfaces/params.interface.ts
export interface PageParams {
	pageNum?: number
	pageSize?: number
	keyword?: string
}
// ...
export interface createUserParams {
	username?: string
	telephone?: string
	realName?: string
	releases?: number
	active?: boolean
	remember?: boolean
	id?: string | number
	roleName?: string
	roleCode?: string
	description?: string
	permissionName?: string
	enabled?: boolean
	menuName?: string
	menuCode?: string
	url?: string
	menuPath?: string
	sort?: number
	level?: number
	type?: number
	parentId?: number | string
	roleId?: number[]
	password?: string
	rePassword?: string
}

export interface rolesListParams {
	remember?: string[]
}

export interface websitUserParams {
	origin?: number
	keyword?: string
}

export interface changeMenusParams extends addMenusParams {
	id?: string
}
export interface addMenusParams {
	parentId?: string
	menuName: string
	url?: string
	menuPath?: string
	sort?: number
	level: number
	leaf?: boolean
	type: number
	description?: string
	icon: string
}

export interface comboParams {
	keyword: string
}

export interface currencyParams {
	keywords: string
	pageNum: number
	pageSize: number
}

export interface rolesParams {
	userId?: string
	roleIds?: string[]
}

export interface permissionsParams {
	roleId: number
	permissionIdList: number[] | string[]
}

export interface savePermissionsParams {
	permissionId: number
	type: number
	resourceIds: ReactText[]
}

export interface searchPermissionsParams {
	permissionId: number
	type: number
}

export interface editPassWordParams {
	id: string
	oldPassword: string
	newPassword: string
	rePassword: string
}
// ...
export interface developersParams {
	pageNum: number
	pageSize: number
	username?: string
	name?: string
	tag?: string
	type?: number
	active?: boolean
}
export interface updateDevParams {
	id: string
	tag?: string
	publishCount?: number
	active?: boolean
}
export interface updateSourceParams {
	devId: string
	devPackageId: string
	expire: number
	publishScenesExtra?: number
	onlineServices?: number
	visitorInfos?: number
	lives?: number
	sceneSyncs?: number
	version: number
	vrLooks?: number
}
// ...
export interface logListParams {
	keyword: string
	creator: string
	startTime: string
	endTime: string
	businessType: string
}
// ...
export interface exportParams {
	params: number[]
	keyword: string
	startTime: string
	endTime: string
}